requirements.txt
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-mgmt-compute==30.5.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
azure-storage-queue==12.9.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.29
botocore==1.34.29
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
dacite==1.8.1
databricks==0.2
databricks-connect==14.3.2
databricks-sdk==0.28.0
databricks-sql-connector==3.1.0
dateutils==0.6.12
decorator==5.1.1
duckdb==0.9.2
executing==2.0.1
extra-streamlit-components==0.1.60
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
gw_dsl_parser==0.1.44
htmlmin==0.1.12
idna==3.6
ImageHash==4.3.1
importlib-metadata==7.0.1
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
#joblib==1.1.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.9
kanaries_track==0.0.4
kiwisolver==1.4.5
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.6
mdurl==0.1.2
missingno==0.5.2
monotonic==1.6
multimethod==1.4
networkx==3.2.1
numba==0.58.1
numpy==1.25.2
oracledb==2.5.0
packaging==23.2
pandas==2.1.4
# pandas-profiling==3.2.0
parso==0.8.3
patsy==0.5.6
pdf2image==1.17.0
phik==0.12.4
pillow==10.2.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pygwalker==0.4.4
PyJWT==2.8.0
pyodbc==5.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.5.0
pyxlsb==1.0.10
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
s3transfer==0.10.0
scipy==1.11.4
seaborn==0.12.2
segment-analytics-python==2.2.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.25
sqlglot==20.10.0
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.30.0
scikit-learn==1.4.0
streamlit-authenticator==0.3.1
#streamlit-pandas-profiling==0.1.3
streamlit-scrollable-textbox==0.0.3
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typeguard==4.1.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
visions==0.7.5
wasmtime==12.0.0
watchdog==3.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.9
wordcloud==1.9.3
xlrd==2.0.1
ydata-profiling==4.6.4
zipp==3.17.0
streamlit-modal==0.1.2
openpyxl==3.1.2
